+++ src/Commands/QueueClearAllCommand.php
@@ -15,9 +15,18 @@ class QueueClearAllCommand extends Command
     {
         $connection = $this->argument('connection') ?? config('queue.default');
 
+        $defaults = collect(config('horizon.defaults', []));
+
         $queues = collect(config('horizon.environments', []))
-            ->flatMap(fn (array $environment): array => collect($environment)->pluck('queue')->all())
+            ->flatMap(fn (array $environment): array => collect($environment)
+                ->map(fn (array $supervisor, string $name): array => array_merge(
+                    $defaults->get($name, []),
+                    $supervisor,
+                ))
+                ->pluck('queue')
+                ->all())
             ->flatten()
+            ->filter()
             ->unique()
             ->values()
             ->all();